requirements.txt
CherryPy==3.2.2
Jinja2==2.6
Mako==0.3.6
MarkupSafe==0.11
MySQL-python==1.2.3
Paver==1.0.4
PyYAML==3.10
SQLAlchemy==0.7.9
WTForms==1.0.2
alembic==0.4.0
argparse==1.2.1
colorlog==1.6
configobj==4.7.2
ensconce==1.3
netaddr==0.7.5
nose==1.2.1
pexpect==2.4
psycopg2==2.4.2
py-bcrypt==0.2
pycrypto==2.6
python-dateutil==1.5
python-gnupg==0.3.1
python-ldap==2.3.11
pytz==2013b
requests==0.14.2
selenium==2.28.0
unittest2==0.5.1